🎨 art: Long -> Integer로 수정
- Board의 id 타입을 Integer로 변경

+++ src/main/java/devholic22/board/controller/BoardController.java
@@ -36,7 +36,7 @@ public class BoardController {
     // 특정 게시물 조회
     @GetMapping("/{id}")
     public String OneBoard(@PathVariable String id, Model model) {
-        Board board = boardService.findOne(Long.parseLong(id));
+        Board board = boardService.findOne(Integer.parseInt(id));
         model.addAttribute("board", board);
         return "board";
     }
@@ -44,7 +44,7 @@ public class BoardController {
     // 특정 게시물 수정 페이지
     @GetMapping("/{id}/edit")
     public String EditBoard(@PathVariable String id, Model model) {
-        Board board = boardService.findOne(Long.parseLong(id));
+        Board board = boardService.findOne(Integer.parseInt(id));
         model.addAttribute("board", board);
         return "editForm";
     }
@@ -53,14 +53,14 @@ public class BoardController {
     // PutMapping이 통하지 않는다.. 조치 완료!
     @PutMapping("/{id}")
     public String EditBoard(@PathVariable String id, @ModelAttribute BoardDto boardDto) {
-        boardService.fix(Long.valueOf(id), boardDto);
+        boardService.fix(Integer.valueOf(id), boardDto);
         return "redirect:/";
     }
 
     // 특정 게시물 삭제 요청
     @DeleteMapping("/{id}")
     public String DeleteBoard(@PathVariable String id) {
-        boardService.remove(Long.valueOf(id));
+        boardService.remove(Integer.valueOf(id));
         return "redirect:/";
     }
 }
